Revert "#125 add posibility to add modifiers when clicking on component"
This reverts commit c6f367085d52a3ec4ccfc5faa94dbbee584fc56c.
