+++ src/main/java/org/robotframework/swing/testapp/TestButton.java
@@ -2,7 +2,6 @@ package org.robotframework.swing.testapp;
 
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
-import java.awt.event.ActionEvent;
 
 import javax.swing.JButton;
 
@@ -10,7 +9,6 @@ public class TestButton extends JButton {
     private static final String INITIAL_TEXT = "Test Button";
     private static final String BUTTON_PUSHED_TEXT = "Button Was Pushed";
     private static final String RIGHT_CLICKED_TEXT = "Right Clicked";
-    private static final String CLICK_SHIFT_TEXT = "Click With Shift";
 
     public TestButton() {
         super(INITIAL_TEXT);
@@ -21,8 +19,6 @@ public class TestButton extends JButton {
             public void mouseClicked(MouseEvent e) {
                 if (e.getButton() == MouseEvent.BUTTON1)
                     setText(BUTTON_PUSHED_TEXT + e.getClickCount());
-                if(e.getButton() == MouseEvent.BUTTON1 && e.isShiftDown())
-                    setText(CLICK_SHIFT_TEXT);
                 if (e.getButton() == MouseEvent.BUTTON3)
                     setText(RIGHT_CLICKED_TEXT);
             }
